requirements.txt
absl-py==2.1.0
alembic==1.16.1
annotated-types==0.7.0
anyio==4.9.0
astunparse==1.6.3
babel==2.17.0
black==25.1.0
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.12
Flask==3.0.3
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
flatbuffers==24.3.25
fonttools==4.54.1
fpdf==1.7.2
gast==0.6.0
google-pasta==0.2.0
greenlet==3.1.1
grpcio==1.68.0
h11==0.14.0
h5py==3.12.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
keras==3.7.0
kiwisolver==1.4.7
libclang==18.1.1
Mako==1.3.10
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mdurl==0.1.2
ml-dtypes==0.4.1
mypy_extensions==1.1.0
namex==0.0.8
numpy==2.0.2
opencv-python==4.10.0.84
opt_einsum==3.4.0
optree==0.13.1
packaging==24.1
pandas==2.2.3
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.3.8
protobuf==5.29.0
psycopg2-binary==2.9.10
pydantic==2.11.1
pydantic_core==2.33.0
pygame==2.5.2
Pygments==2.18.0
pymongo==4.11.3
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
rich==13.9.4
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
setuptools==75.6.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow_intel==2.18.0
termcolor==2.5.0
threadpoolctl==3.5.0
tkcalendar==1.6.1
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.34.0
Werkzeug==3.0.4
wheel==0.45.1
wrapt==1.17.0
WTForms==3.2.1
